Move key definition inline with its service
